requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
tqdm>=4.60.0
requests>=2.25.0
reportlab>=3.6.0
anthropic>=0.25.0
python-dotenv>=0.19.0 